+++ openaq/core/transport.py
@@ -1,7 +1,5 @@
 """Base class and utility functions for working with client transport."""
 
-from __future__ import annotations
-
 import http.client
 import json
 import logging
@@ -13,7 +11,7 @@ from collections import deque
 from collections.abc import Mapping
 from dataclasses import dataclass, field
 from http import HTTPStatus
-from typing import Any
+from typing import Any, Self
 
 from openaq.core.exceptions import (
     BadGatewayError,
@@ -117,15 +115,13 @@ class Headers(dict[str, str]):
         for k, v in kwargs.items():
             self[k] = v
 
-    def copy(self) -> Headers:
+    def copy(self) -> Self:
         """Returns a shallow copy of this Headers instance.
 
         Returns:
             A new Headers instance with the same key-value pairs.
         """
-        h = Headers()
-        super(Headers, h).update(self)
-        return h
+        return type(self)(self)
 
 
 class Response:
@@ -534,7 +530,7 @@ class Transport:
         self._pool.close_all()
 
 
-_HTTP_SATUS_MAP = {
+_HTTP_STATUS_MAP = {
     HTTPStatus.BAD_REQUEST: BadRequestError,
     HTTPStatus.NOT_FOUND: NotFoundError,
     HTTPStatus.REQUEST_TIMEOUT: TimeoutError,
@@ -583,7 +579,7 @@ def check_response(res: Response) -> Response:
     except ValueError:
         http_status = None
     exc_class = (
-        _HTTP_SATUS_MAP.get(http_status, ServerError) if http_status else ServerError
+        _HTTP_STATUS_MAP.get(http_status, ServerError) if http_status else ServerError
     )
     logger.error("HTTP %s - %s", res.status_code, res.text)
     raise exc_class(res.text)