+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ###
 
-###
+### Changed
+
+- Removed conditional import for `tomllib` based on >=3.11 support.
 
 
 ## [1.1.0] - 2026-07-02